EDSC-3966: rename util/metaIcon.js to getCustomizeIcons.js
