+++ static/src/js/util/getCustomizeIcons.js
